[config] Add config file support
Create a Config class to hold configuration set in a file and also command-line
arguments.  It reads config items from a hierarchy of files, with later reads
overwriting values from earlier reads.  The read order is:

    /etc/dartt/config.toml
    $HOME/.dartt/config.toml
    $HOME/.config/dartt/config.toml

If no config file is found, the user is asked if they want to create one.